refactor(regime): split scorers out of cri_scoring; preserve import surface
cri_scoring.py reached 504 lines after the v3 changes — at the soft budget.
Extracted constants + scorers + composite + CTA + crash to cri_scorers.py;
cri_scoring.py now holds the run_analysis orchestrator and re-exports
every name for back-compat. 44/44 unit tests still pass.
